Add NumAsHexProtocol; Inherit from it: Felt, NumAsHex
